+++ src/jobs/UpdateProductMetadata.php
@@ -2,10 +2,11 @@
 
 namespace craft\shopify\jobs;
 
-use Craft;
 use craft\queue\BaseJob;
 use craft\shopify\elements\Product;
+use craft\shopify\helpers\Metafields as MetafieldsHelper;
 use craft\shopify\Plugin;
+use craft\shopify\records\ProductData as ProductDataRecord;
 
 /**
  * Updates the metadata for a Shopify product.
@@ -20,11 +21,15 @@ class UpdateProductMetadata extends BaseJob
     public function execute($queue): void
     {
         $api = Plugin::getInstance()->getApi();
-        $product = Product::find()->shopifyId($this->shopifyProductId)->one();
-        if ($product) {
-            $metaFields = $api->getMetafieldsByProductId($this->shopifyProductId);
+
+        if ($product = Product::find()->shopifyId($this->shopifyProductId)->one()) {
+            $metaFieldsObjects = $api->getMetafieldsByProductId($this->shopifyProductId);
+            $metaFields = MetafieldsHelper::unpack($metaFieldsObjects);
             $product->setMetafields($metaFields);
-            Craft::$app->elements->saveElement($product);
+            /** @var ProductDataRecord $productData */
+            $productData = ProductDataRecord::find()->where(['shopifyId' => $this->shopifyProductId])->one();
+            $productData->metaFields = $metaFields;
+            $productData->save();
             sleep(1); // Avoid rate limiting
         }
     }