+++ src/records/ProductData.php
@@ -33,6 +33,7 @@ use yii\db\ActiveQueryInterface;
  * @property string $updatedAt
  * @property array $variants
  * @property string $vendor
+ * @property array $metaFields
  * @property string $dateUpdated
  *
  */